+++ NEWS.md
@@ -8,6 +8,9 @@
 * `settings$package.dependency.fields()` now only affects packages installed
   directly by the user, not downstream dependencies of those packages.
 
+* Fixed an issue where `renv::snapshot(exclude = <...>)` could warn when
+  attempting to exclude a package which was not already installed. (#1396)
+
 * If `renv::snapshot()` finds missing packages, a new prompt allows you to 
   install them before continuing (#1198).